chore: refactor token list and add tests
